+++ authentication-service/src/main/kotlin/com/saveourtool/save/authservice/security/ConvertingAuthenticationManager.kt
@@ -31,9 +31,7 @@ class ConvertingAuthenticationManager(
      * @throws BadCredentialsException in case of bad credentials
      */
     override fun authenticate(authentication: Authentication): Mono<Authentication> = if (authentication is UsernamePasswordAuthenticationToken) {
-        println("\n\n\n-------------------authenticate authservice!!!!")
         val (name, identitySource) = authentication.extractUserNameAndIdentitySource()
-        println("name: $name identitySource $identitySource")
         userDetailsService.findByUsername(name)
             .cast<IdentitySourceAwareUserDetails>()
             .filter {
